# Generalize `rrule` for `svdvals`