Fixes #206.
Fixes #587.