@devmotion 2024-04-16
A simple comparison:

## main

```julia
julia> using ChainRules, BenchmarkTools, LinearAlgebra

julia> @btime ChainRules.rrule(svdvals, $(Symmetric(randn(50, 50))));
  155.542 μs (17 allocations: 79.92 KiB)

julia> @btime $(ChainRules.rrule(svdvals, Symmetric(randn(50, 50)))[2])($(randn(50)));
  7.903 μs (7 allocations: 40.34 KiB)
```

## This PR

```julia
julia> @btime ChainRules.rrule(svdvals, $(Symmetric(randn(50, 50))));
  155.875 μs (20 allocations: 118.78 KiB)

julia> @btime $(ChainRules.rrule(svdvals, Symmetric(randn(50, 50)))[2])($(randn(50)));
  7.597 μs (6 allocations: 59.02 KiB)
```

## Setup

```julia
julia> versioninfo()
Julia Version 1.10.2
Commit bd47eca2c8a (2024-03-01 10:14 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: macOS (arm64-apple-darwin22.4.0)
  CPU: 10 × Apple M2 Pro
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-15.0.7 (ORCJIT, apple-m1)
Threads: 1 default, 0 interactive, 1 GC (on 6 virtual cores)
Environment:
  JULIA_PKG_USE_CLI_GIT = true
```